+++ components/corrected/HearingTask.vue
@@ -16,7 +16,7 @@
         </div>
       </div>
       <div v-show="helpOpen" class="px-4 py-5 border-b border-gray-200 sm:px-6">
-        Click on play, guess the MARVEL movie and write down your answer as fast as you can.
+        <b>Fehler: </b> Ihre Antwort ist leider falsch!
       </div>
       <div class="px-4 py-5">
         <input type="hidden" :name="'hearingtask-'+props.index+'-id'" :value="props.id+','+input">
@@ -44,24 +44,24 @@
         </div>
         <div class="grid grid-cols-5 cols-gap-1.5">
           <div style="padding:10px;">
-            <input id="answerI" v-model="input" type="checkbox" value="answer1" class="mr-2">
-            <label id="aILabel" for="answerI" :style="[!props.fill ? {color: 'black'}: answersCorrect.aCI ? {color:'green'}:{color:'red'}]">{{ answerI }}</label>
+            <input id="answerI" v-model="inputI" type="checkbox" value="answer1" class="mr-2">
+            <label id="aILabel" for="answerI" :style="[answersCorrect.aCI ? {color:'green'}:{color:'red'}]">{{ answerI }}</label>
           </div>
           <div style="padding:10px;">
-            <input id="answerII" v-model="input" type="checkbox" value="answer2" class="mr-2">
-            <label id="aIILabel" for="answerII" :style="[!props.fill ? {color: 'black'}: answersCorrect.aCII ? {color:'green'}:{color:'red'}]">{{ answerII }}</label>
+            <input id="answerII" v-model="inputII" type="checkbox" value="answer2" class="mr-2">
+            <label id="aIILabel" for="answerII" :style="[answersCorrect.aCII ? {color:'green'}:{color:'red'}]">{{ answerII }}</label>
           </div>
           <div style="padding:10px;" >
-            <input id="answerIII" v-model="input" type="checkbox" value="answer3" class="mr-2">
-            <label id="aIIILabel" for="answerIII" :style="[!props.fill ? {color: 'black'}: answersCorrect.aCIII ? {color:'green'}:{color:'red'}]">{{ answerIII }}</label>
+            <input id="answerIII" v-model="inputIII" type="checkbox" value="answer3" class="mr-2">
+            <label id="aIIILabel" for="answerIII" :style="[answersCorrect.aCIII ? {color:'green'}:{color:'red'}]">{{ answerIII }}</label>
           </div>
           <div style="padding:10px;">
-            <input id="answerIV" v-model="input" type="checkbox" value="answer4" class="mr-2">
-            <label id="aIVLabel" for="answerIV" :style="[!props.fill ? {color: 'black'}: answersCorrect.aCIV ? {color:'green'}:{color:'red'}]">{{ answerIV }}</label>
+            <input id="answerIV" v-model="inputIV" type="checkbox" value="answer4" class="mr-2">
+            <label id="aIVLabel" for="answerIV" :style="[answersCorrect.aCIV ? {color:'green'}:{color:'red'}]">{{ answerIV }}</label>
           </div>
           <div style="padding:10px;">
-            <input id="answerV" v-model="input" type="checkbox" value="answer5" class="mr-2">
-            <label id="aVLabel" for="answerV" :style="[!props.fill ? {color: 'black'}: answersCorrect.aCV  ? {color:'green'}:{color:'red'}]">{{ answerV }}</label>
+            <input id="answerV" v-model="inputV" type="checkbox" value="answer5" class="mr-2">
+            <label id="aVLabel" for="answerV" :style="[answersCorrect.aCV  ? {color:'green'}:{color:'red'}]">{{ answerV }}</label>
           </div>
         </div>
       </div>
@@ -90,6 +90,10 @@ const props = defineProps({
   fillElements: {
     type: String,
     required: true
+  },
+  correct: {
+    type: String,
+    required: true
   }
 });
 const url = "/api/hearingTask/" + props.id;
@@ -97,31 +101,18 @@ const { data: api } = await useFetch(url, {
   key: hash([url])
 });
 const question = api.value.question;
+const answersCorrectList = props.correct.split(",");
 const answers = api.value.answers.split(",");
-const answersCorrectList = api.value.answersCorrect.split(",");
-const answersCorrect = { aCI: answersCorrectList[0] === "true", aCII: answersCorrectList[1] === "true", aCIII: answersCorrectList[2] === "true", aCIV: answersCorrectList[3] === "true" };
 const answerI = answers[0];
 const answerII = answers[1];
 const answerIII = answers[2];
 const answerIV = answers[3];
 const answerV = answers[4];
-// const input = reactive("");
-const answersGiven = reactive({ aGI: false, aGII: false, aGIII: false, aGIV: false, aGV: false });
-if (props.fill) {
-  answersGiven.aGI = (props.fillElements.includes("answer1") && answersCorrect.aCI) || (!props.fillElements.includes("answer1") && !answersCorrect.aCI);
-  answersGiven.aGII = (props.fillElements.includes("answer2") && answersCorrect.aCII) || (!props.fillElements.includes("answer2") && !answersCorrect.aCII);
-  answersGiven.aGIII = (props.fillElements.includes("answer3") && answersCorrect.aCIII) || (!props.fillElements.includes("answer3") && !answersCorrect.aCIII);
-  answersGiven.aGIV = (props.fillElements.includes("answer4") && answersCorrect.aCIV) || (!props.fillElements.includes("answer4") && !answersCorrect.aCIV);
-  answersGiven.aGV = (props.fillElements.includes("answer4") && answersCorrect.aCV) || (!props.fillElements.includes("answer4") && !answersCorrect.aCV);
-}
-const helpOpen = ref(false);
-</script>
-<script>
-export default {
-  data () {
-    return {
-      input: []
-    };
-  }
-};
+const answersCorrect = { aCI: answersCorrectList[0] === "true", aCII: answersCorrectList[1] === "true", aCIII: answersCorrectList[2] === "true", aCIV: answersCorrectList[3] === "true", aCV: answersCorrectList[4] === "true" };
+const inputI = (props.fillElements.includes("answer1") ? "true" : "false");
+const inputII = (props.fillElements.includes("answer2") ? "true" : "false");
+const inputIII = (props.fillElements.includes("answer3") ? "true" : "false");
+const inputIV = (props.fillElements.includes("answer4") ? "true" : "false");
+const inputV = (props.fillElements.includes("answer5") ? "true" : "false");
+const helpOpen = ref(answersCorrectList.includes("false"));
 </script>